Add files via upload
California is the first state to  banned the sale of new gasoline vehicles in 2035. My analysis focuses on projecting the sales and market share of electric vehicles as well as it provides insights into the trend and adoption of gas powered vehicles in California by 2035. So, lets see if it is really feasible for California to enforce the ban.